Update deleteNote mutation - authorization
- Add userId parameter to delete method on notes datasource
- Update resolver to use the user id from the auth token when
  deleting a note, thus preventing unauthorized access. A user can only
  delete their own notes.
- Add test to api.test to check for an unauthorized attempt.

+++ server/datasources/Notes.source.js
@@ -28,12 +28,12 @@ class NotesSource extends MongoDataSource {
   }
 
   /**
-   * Delete a note for given noteId.
+   * Delete a note for given noteId and userId.
    *
    * Returns a promise which resolves 1 if note is deleted and 0 if no matching note is found.
    */
-  async delete({ noteId }) {
-    const result = await this.model.deleteOne({ _id: noteId });
+  async delete({ noteId, userId }) {
+    const result = await this.model.deleteOne({ _id: noteId, user: userId });
     return result.deletedCount;
   }
 }
